requirements.txt
faiss-cpu==1.7.2
fastapi==0.80.0
joblib==1.2.0
lmdb==1.3.0
numba==0.56.4
numpy>1.23.2
Pillow==9.2.0
pydantic==1.9.2
scipy==1.9.0
tqdm==4.64.0
uvicorn==0.18.3
python-multipart
